+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## Unreleased
+
+- bug fix in `Account.QrCode()`
+
 ## v0.5.0
 
-- Account.QrCode() now returns `(string, string, error)` instead of `([2]string, err)`
+- `Account.QrCode()` now returns `(string, string, error)` instead of `([2]string, err)`
 
 ## v0.4.0